requirements.txt
streamlit~=1.41.1
tensorflow~=2.18.0
requests~=2.32.3
numpy~=1.26.0
pillow~=11.1.0
gdown~=5.2.0
keras~=3.8.0